Moved LAPACKLibrary into linalg, started setting up CodeLAPACK
